refactor: finish Access repository boundaries (#370)
