+++ internal/access/postgres/canonical_credentials.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"github.com/google/uuid"
