+++ internal/access/postgres/directory_identity_helpers.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"database/sql"
