+++ internal/access/postgres/directory_identity_reconcile_repository.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"context"
