+++ internal/access/postgres/sso_oauth_provider_write.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"fmt"
