+++ internal/access/postgres/sso_repository_scan.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"bytes"
