+++ internal/repository/canonical_credentials_integration.e2e
@@ -213,6 +213,55 @@ func TestDeleteForTeamWithAuditCommitsBeforePrivateErasureQueue(t *testing.T) {
 		return tx.Raw(`SELECT status FROM credentials WHERE id = ?`, rollbackTarget.ID).Row().Scan(&status)
 	}))
 	require.Equal(t, "active", status)
+
+	t.Run("credential-private queue failure rolls back credential membership audit and queue", func(t *testing.T) {
+		queueFailureTarget := createOwnedCredential(t, credentialRepo, teamID, ownerID, "audit-queue-rollback", domain.CredentialBindingCredentialPrivate)
+		require.NoError(t, rls.WithSystemTx(ctx, adminDB, func(tx *gorm.DB) error {
+			return tx.Exec(`
+				UPDATE memory_spaces
+				SET lifecycle_state = 'retired', retired_at = NOW()
+				WHERE id = ? AND team_id = ? AND owner_credential_id = ?
+			`, queueFailureTarget.MemorySpaceID, teamID, queueFailureTarget.ID).Error
+		}))
+
+		_, err := credentialRepo.DeleteForTeamWithAudit(ctx, teamID, queueFailureTarget.ID, CredentialDeletionAuditInput{
+			ActorRole:     "manager",
+			ClientIP:      "127.0.0.1",
+			CorrelationID: "corr-delete-queue-rollback",
+		})
+		require.Error(t, err)
+
+		var status, membershipStatus string
+		var auditCount, queuedCount int64
+		require.NoError(t, rls.WithSystemTx(ctx, adminDB, func(tx *gorm.DB) error {
+			if err := tx.Raw(`SELECT status FROM credentials WHERE id = ?`, queueFailureTarget.ID).Row().Scan(&status); err != nil {
+				return err
+			}
+			if err := tx.Raw(`
+				SELECT status
+				FROM team_memberships
+				WHERE team_id = ? AND actor_identity_id = ?
+			`, teamID, queueFailureTarget.ID).Row().Scan(&membershipStatus); err != nil {
+				return err
+			}
+			if err := tx.Raw(`
+				SELECT COUNT(*)
+				FROM audit_log
+				WHERE team_id = ? AND entity_type = 'api_key' AND entity_id = ? AND operation = 'DELETE'
+			`, teamID, queueFailureTarget.ID.String()).Row().Scan(&auditCount); err != nil {
+				return err
+			}
+			return tx.Raw(`
+				SELECT COUNT(*)
+				FROM private_memory_erasure_operations
+				WHERE team_id = ? AND target_credential_id = ?
+			`, teamID, queueFailureTarget.ID).Row().Scan(&queuedCount)
+		}))
+		require.Equal(t, "active", status)
+		require.Equal(t, "active", membershipStatus)
+		require.Zero(t, auditCount)
+		require.Zero(t, queuedCount)
+	})
 }
 
 func createLedgerSSOIdentity(t *testing.T, db *gorm.DB, rls *storagepostgres.RLS, teamID uuid.UUID) uuid.UUID {
