+++ internal/repository/directory_identity_repository_integration.e2e
@@ -89,7 +89,7 @@ func TestDirectoryReconcileCreatesAndArchivesOnlyDirectoryManagedTeam(t *testing
 	profiles, err := ssoRepo.ListTeamMembershipsForIdentity(ctx, user.IdentityID)
 	require.NoError(t, err)
 	require.Len(t, profiles, 1)
-	require.Equal(t, directoryMembershipName(user.Email, user.DisplayName, user.IdentityID), profiles[0].Membership.Name)
+	require.Equal(t, user.DisplayName+" ("+user.IdentityID.String()+")", profiles[0].Membership.Name)
 	entitled, err := ssoRepo.DirectoryMembershipEntitled(ctx, provider.ID, user.IdentityID, teamID, group.ExternalID)
 	require.NoError(t, err)
 	require.True(t, entitled)
